fix(python): resolve remaining ruff lint + format findings post-rebase
Address the lint errors that caused Platform Tests to fail (CI run 79654741125):
- C420: replace `{pid: True for pid in ids}` with `dict.fromkeys(ids, True)`
- RUF059: prefix unused unpacked tuple variables with `_`
- UP006 / UP035: `typing.List` -> `list` in tests/test_trigger_platform_plumbing.py
- TID252: switch local `from ..triggers.service` import to absolute path in
  src/plugins/base.py (keep the lazy-import pattern intact to avoid the
  TriggerService <-> PluginBase cycle)
- Reformat 4 files (base.py, manifest.py, test_plugin_validation.py,
  test_trigger_platform_plumbing.py) and one test_plugin_registry.py block
  that ruff format wanted to rewrap

Also:
- RUF013 / int(round(...)) cleanup carried over from initial sweep
- Use spread syntax for message-list concat in src/ai/chat.py
- Tighten `pytest.raises(... match=...)` to raw string in tests/ai/test_chat.py
- Allow RUF059 in src/api_server.py per-file-ignores (matches the existing
  CodeQL exemption strategy for that file)

+++ src/utils/surf.py
@@ -240,7 +240,7 @@ class SurfSource:
             Formatted message: "OB SURF: [Height]ft @ [Period]s"
         """
         # Round period to integer for cleaner display
-        period_display = int(round(swell_period)) if swell_period else 0
+        period_display = round(swell_period) if swell_period else 0
         return f"OB SURF: {wave_height}ft @ {period_display}s"
